Adding taxes in the database and UI (#13)
Implementing the tax feature in the database and UI. The following assumptions were made

There is a single tax rate
There is a flag in the Products table that marks whether taxes apply or not
This was done to keep things simple. We could add a Tax table that keeps a tab on different tax rates. Probably this can be an enhancement for the future.